Fixed seeking issues on some DVDs where, for example, trying to jump backwards could actually cause playback to jump forwards.
dvdnav_absolute_time_search in libdvdnav (renamed from the original dvdnav_time_search but functionally the same) uses a time offset with a start sector, which will only work if all NAV packets are equally spaced.  On at least one DVD that had issues, the distance between consecutive NAV packets ranged from 96 to 457.

The existing mechanism to check whether a seek was successful and, if not, narrow in on the required time was broken when jumping backwards.


Fixes #11572

+++ mythtv/libs/libmythtv/DVD/dvdringbuffer.h
@@ -128,7 +128,7 @@ class MTV_PUBLIC DVDRingBuffer : public RingBuffer
 
     virtual void IgnoreWaitStates(bool ignore) { m_skipstillorwait = ignore; }
     void AudioStreamsChanged(bool change) { m_audioStreamsChanged = change; }
-    uint GetCurrentTime(void)             { return (m_currentTime / 90000); }
+    int64_t GetCurrentTime(void)          { return (m_currentTime / 90000); }
     uint TitleTimeLeft(void);
     void  SetTrack(uint type, int trackNo);
     int   GetTrack(uint type);
@@ -197,8 +197,8 @@ class MTV_PUBLIC DVDRingBuffer : public RingBuffer
     const char    *m_dvdname;
     const char    *m_serialnumber;
     bool           m_seeking;
-    uint64_t       m_seektime;
-    uint           m_currentTime;
+    int64_t        m_seektime;
+    int64_t        m_currentTime;
     QMap<uint, uint> m_seekSpeedMap;
     QMap<uint, QList<uint64_t> > m_chapterMap;
